Extract sort functions from GuestsList component
The sort functions could be useful for other components, move them into
a separate utils file for easier reuse and better GuestsList
readability.
